Add files via upload
9-01-10: Works with actual servos. Cleaned up and simplified, both PLAY and RECORD work well.
		Tested with USB ROBOTNIK and ROBOTNIK BRAIN BOARD running four servo motors.

+++ SerialCtrlDemo/SerialCtrl.cpp
@@ -39,7 +39,7 @@ HANDLE SerialCtrl::GetPortHandle(void)
 
 
 
-BOOL SerialCtrl::OpenPort(const char * portName, UINT16 baudrate)
+BOOL SerialCtrl::OpenPort(const char * portName, int baudrate)
 {
 	CString strPortName;
 	strPortName.Format("\\\\.\\%s", portName);
